Add test functions for pressure constraint

+++ include/loads_bcs/constraints.h
@@ -55,7 +55,7 @@ class Constraints {
   //! \param[in] pconstraint Pressure constraint at node
   bool assign_nodal_pressure_constraint(
       const std::shared_ptr<FunctionBase>& mfunction, int set_id,
-      const unsigned phase, const unsigned pconstraint);
+      const unsigned phase, const double pconstraint);
 
   //! Assign nodal pressure constraints to nodes
   //! \param[in] pressure_constraints Constraint at node, pressure
